+++ web/src/reducers/nav.items.js
@@ -110,8 +110,8 @@ const buildActivitySection = i => [
     url: `/apd/activity/${i}/overview`
   },
   {
-    label: 'Objectives and milestones',
-    url: `/apd/activity/${i}/okrs`
+    label: 'Outcomes and milestones',
+    url: `/apd/activity/${i}/oms`
   },
   {
     label: 'State cost categories',
@@ -160,7 +160,7 @@ const setSelected = (items, value, parents = []) => {
 
 const getItems = ({ activities = [], items = staticItems, url = '' } = {}) => {
   // eslint-disable-next-line no-param-reassign
-  items = copy(items);  // copy items state so we don't modify original state
+  items = copy(items); // copy items state so we don't modify original state
   if (activities.length) {
     const item = items.find(i => i.label === 'Activities');
     item.items = [