+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Anticipated release: October 26th, 2020
 #### 🚀 New features
 
 - Update Program Activities headers, labels, and help text ([#2528])
+- Rename Objectives and Key Results to Outcomes and Metrics ([#2584])
 
 #### 🐛 Bugs fixed
 
@@ -15,3 +16,4 @@ Anticipated release: October 26th, 2020
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2528]: https://github.com/CMSgov/eAPD/issues/2528
+[#2584]: https://github.com/CMSgov/eAPD/issues/2584